requirements.txt
coverage==4.3.1
Django==1.10.5
djangorestframework==3.5.3
